refactor: UserCardColumn component VAC pattern 적용
UserCardColumn 컴포넌트에는 관심사를 비즈니스 로직에 두고 VAC component인 UserCardColumnView에 관심사를 UI만 두게 설정.
